+++ src/types/award.ts
@@ -1,17 +1,9 @@
 import { EntityBase } from "./entityBase";
 import { HrefEntity } from "./hrefEntity";
-import { UserBasicEntity } from "./userBasicEntity";
-
-export interface YearEntity extends EntityBase<number> {
-    guestCount: number;
-    description: string;
-    photoId?: number;
-    photo?: HrefEntity;
-    hardhead: UserBasicEntity;
-}
+import { YearEntity } from "./yearEntity";
 
 export interface Award extends EntityBase<number> {
-    winners: HrefEntity;
-    href?: string;
-    years: YearEntity[];
+  winners: HrefEntity;
+  href?: string;
+  years: YearEntity[];
 }
